NewsCollector/requirements.txt
Django==3.2.21
Django==5.2
djangorestframework==3.16.0
